+++ app/src/Abstractions/Properties/BehaviorEntryAbstract.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace Solis\PhpSchema\Abstractions\Properties;
+
+/**
+ * Class BehaviorEntryAbstract
+ *
+ * @package Solis\PhpSchema\Abstractions\Properties
+ */
+abstract class BehaviorEntryAbstract
+{
+
+    /**
+     * @var boolean
+     */
+    protected $required;
+
+    /**
+     * @return bool
+     */
+    public function isRequired()
+    {
+        return $this->required;
+    }
+
+    /**
+     * @param bool $required
+     */
+    public function setRequired($required)
+    {
+        $this->required = $required;
+    }
+
+    /**
+     * toArray
+     */
+    public function toArray()
+    {
+        $array = [
+            'required' => $this->isRequired()
+        ];
+        return $array;
+    }
+}