+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All the relevant changes for this project will be documented following the [keeping changelog](http://keepachangelog.com/) especifications
 
+## 0.1.0 - 2017-06-21
+- Added behavior entry properties schema group
+- Added actions entry in database schema group 
+
+### Added
+- Implemented semantic versioning [Semantic Versioning 2.0.0](http://semver.org/)
+
 ## 0.0.1 - 2017-06-12
 
 ### Added